news/requirements.txt
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
iniconfig==2.0.0
mysql-connector-python==8.1.0
packaging==23.1
pluggy==1.3.0
protobuf==4.21.12
pytest==7.4.2
python-decouple==3.8
python-dotenv==1.0.0
requests==2.31.0
soupsieve==2.5
urllib3==2.0.5
